app/requirements.txt
fastapi
uvicorn
psycopg2
python-dotenv
jinja2
python-multipart
numpy
argon2-cffi